+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Migrate to shared GA workflows. Refs UIOR-1380.
 * Add setting options for number gernerator - accessing number equals call number. Refs UIOR-1383.
 * PO numbers increase by 2 instead of 1. Refs UIOR-1374.
+* Enhancement help text on Settings > Organizations > Number generator options. Refs UIOR-1384.
 
 ## [7.0.4](https://github.com/folio-org/ui-orders/tree/v7.0.4) (2024-12-31)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v7.0.3...v7.0.4)